add new functions

+++ app/Http/Controllers/ContactsController.php
@@ -1,23 +1,37 @@
 <?php
 
 namespace App\Http\Controllers;
+use Illuminate\Http\Request;
+use App\Models\AccountSendEmail;
+use App\Models\AddressRelation;
+use App\Models\AddressProvider;
+use App\Models\FileReations;
+use App\Models\Account;
+use App\Models\Contact;
+use App\Models\Company;
+use App\Models\Address;
+use App\Models\Country;
+use App\Models\Files;
+use App\Models\Notes;
 use App\Models\User;
 use Auth;
-use Illuminate\Http\Request;
-use App\Models\Company;
-use App\Models\Contact;
 
 class ContactsController extends Controller
 {
-    public function index()
-    {
+    public function index(){
         $id = Auth::user()->id;
-
-
         return view('user.contact.contacts', [
             'contacts' => Contact::where('user_id', $id)->get(),
-            'all_companies' =>Company::all( 'id', 'name' ),
-            'users'=>User::where('id', '!=', Auth::user()->id)->get(['id', 'name', 'email']),
+            'accounts' =>Account::all( 'id', 'name' ),
+            'users'=>User::where('id', '!=', Auth::user()->id)->get(['id','first_name', 'last_name', 'email']),
+        ]);
+    }
+
+    public function contacts_by_account($id){
+        return view('user.contact.contacts', [
+            'contacts' => Contact::where('account_id', $id)->get(),
+            'accounts' =>Account::all( 'id', 'name' ),
+            'users'=>User::where('id', '!=', Auth::user()->id)->get(['id','first_name', 'last_name', 'email']),
         ]);
     }
 
@@ -28,7 +42,7 @@ class ContactsController extends Controller
         $data->user_id = Auth::user()->id;
         $data->owner_id = $req->input('owner_id');
         $data->solution = $req->input('solution');
-        $data->company_id = $req->input('company_id');
+        $data->account_id = $req->input('account_id');
         $data->first_name = $req->input('first_name');
         $data->last_name = $req->input('last_name');
         $data->middle_name = $req->input('middle_name');
@@ -48,14 +62,17 @@ class ContactsController extends Controller
         $data->mailing__zip_code = $req->input('mailing__zip_code');
 
         if($data->save()){
-            return redirect()->route('contacts')->with('success', $req->input('title').' - Add');
+            return redirect()->route('contacts')->with('success', $req->input('title').' - Added');
         }
-
 
     }
 
     public function edit_contact(Request $req, $id){
         $contact = Contact::find($id);
+
+        if(empty($contact)){
+            return redirect()->route('contacts')->with('danger', "Not Found");
+        }
 
         if ($_POST){
             if($contact->user_id == Auth::user()->id) {
@@ -63,7 +80,7 @@ class ContactsController extends Controller
                 $contact->user_id = Auth::user()->id;
                 $contact->owner_id = $req->input('owner_id');
                 $contact->solution = $req->input('solution') !== "0" ? $req->input('solution') : 0 ;
-                $contact->company_id = $req->input('company_id');
+                $contact->account_id = $req->input('account_id');
                 $contact->first_name = $req->input('first_name');
                 $contact->last_name = $req->input('last_name');
                 $contact->middle_name = $req->input('middle_name');
@@ -83,25 +100,50 @@ class ContactsController extends Controller
                 $contact->mailing__zip_code = $req->input('mailing__zip_code');
 
                 if ($contact->save()) {
-                    return redirect()->route('edit_contact', [$id])->with('success', $req->input('title') . ' - Edit');
+                    return redirect()->route('edit_contact', [$id])->with('success', $req->input('title') . ' - Edited');
                 }
             }
         }
+        $notifications = new NotificationController;
 
         return view('user.contact.edit_contact', [
-//            'company_types' => CompanyTypes::all(),
-//            'industries_types' => IndustriesTypes::all(),
             'contact' => $contact,
-            'all_companies' =>Company::all(['id', 'name']),
-            'all_contacts' =>Contact::where('id', '!=', $id)->get(['id', 'title']),
-            'users'=>User::where('id', '!=', Auth::user()->id)->get(['id', 'name','email']),
+            'accounts' => Account::where('user_id', '=', Auth::user()->id)->get(['id', 'name']),
+            'companies' =>Company::where('user_id', '=', Auth::user()->id)->get(['id', 'name']),
+            'contacts' => Contact::where('user_id', '=', Auth::user()->id)->get(['id', 'title']),
+            'users'=> User::where('id', '!=', Auth::user()->id)->get(['id', 'first_name', 'last_name','email']),
+            'emails' => AccountSendEmail::where('user_id', '=', Auth::user()->id)->get(),
+            'notifications'=> $notifications->notifications('contact_id',$id),
+            'companies_count' => Company::where('user_id', '=', Auth::user()->id)->where('contact_id', '=', $id)->get(['id', 'name']),
+            'contacts_count' => Contact::where('user_id', '=', Auth::user()->id)->get(['id', 'title']),
+            'upcoming_overdues' => $notifications->UpcomingOverdue('contact_id',$id),
+            'subject_events' => [1 => 'Call', 2 => 'Email', 3 => 'Meeting', 4 => 'Send Letter/Quote', 5 => 'Other'],
+            'subject_tasks' => [1 => 'Call', 2 => 'Send Letter', 3 => 'Send Quote', 4 => 'Other'],
+            'subject_calls' => [1 => 'Call', 2 => 'Send Letter', 3 => 'Send Quote', 4 => 'Other'],
+            'url' => 'contact',
+            'id' => $id,
+            'page_title' => $contact->title,
+            'notes' => Notes::where('contact_id', '=', $id)->get(),
+            'files' => FileReations::where('contact_id', '=', $id)->where('status', '=', 1)->with('file')->get(),
+            'files_data' => FileReations::where('user_id', '=', Auth::user()->id)->with('file')->get(),
+            'countries' => Country::all(),
+            'address_providers' => AddressProvider::where('user_id', '=', Auth::user()->id)->get(),
+            'addresses' => Address::where('user_id', '=', Auth::user()->id)
+            ->with('country')
+            ->with('state')
+            ->with('addressRelation')
+            ->whereHas('addressRelation', function($q) use($id){$q->where('contact_id', $id);})->get(),
+            'all_addresses' => Address::where('user_id', '=', Auth::user()->id)->with('country')->with('state')->with('addressRelation')->get(),
         ]);
     }
 
     public function delete_contact($id){
         $data = Contact::find($id);
+        if(empty($data)){
+            return redirect()->route('contacts')->with('danger', "Not Found");
+        }
         if($data->delete()){
-            return redirect()->route('contacts')->with('success', $data->title.' - Remove');
+            return redirect()->route('contacts')->with('success', $data->title.' - Removed');
         }
     }
 }
